refactor: split file and import by main
