+++ getting_user_input.py
@@ -1,0 +1,71 @@
+import flet as ft
+
+
+def counter_app(page: ft.Page):
+    page.title = "Flet counter example"
+    page.vertical_alignment = ft.MainAxisAlignment.CENTER
+
+    txt_number = ft.TextField(value="0", text_align=ft.TextAlign.RIGHT, width=100)
+
+    def minus_click(e):
+        txt_number.value = str(int(txt_number.value) - 1)
+        page.update()
+
+    def plus_click(e):
+        txt_number.value = str(int(txt_number.value) + 1)
+        page.update()
+
+    page.add(
+        ft.Row(
+            [
+                ft.IconButton(ft.icons.REMOVE, on_click=minus_click),
+                txt_number,
+                ft.IconButton(ft.icons.ADD, on_click=plus_click),
+            ],
+            alignment=ft.MainAxisAlignment.CENTER,
+        )
+    )
+
+
+def text_box(page: ft.Page):
+    def btn_click(e):
+        if not txt_name.value:
+            txt_name.error_text = "Please enter your name"
+            page.update()
+        else:
+            name = txt_name.value
+            page.clean()
+            page.add(ft.Text(f"Hello, {name}!"))
+
+    txt_name = ft.TextField(label="Your name")
+    page.add(txt_name, ft.ElevatedButton("Say hello!", on_click=btn_click))
+
+
+def check_box(page: ft.Page):
+    def checkbox_changed(e):
+        output_text.value = f"You have learned how to ski : {todo_check.value}."
+        page.update()
+
+    output_text = ft.Text()
+    todo_check = ft.Checkbox(
+        label="ToDo: Learn how to use ski", value=False, on_change=checkbox_changed
+    )
+    page.add(todo_check, output_text)
+
+
+def dropdown(page: ft.Page):
+    def button_clicked(e):
+        output_text.value = f"Dropdown value is: {color_dropdown.value}"
+        page.update()
+
+    output_text = ft.Text()
+    submit_btn = ft.ElevatedButton(text="Submit", on_click=button_clicked)
+    color_dropdown = ft.Dropdown(
+        width=100,
+        options=[
+            ft.dropdown.Option("Red"),
+            ft.dropdown.Option("Green"),
+            ft.dropdown.Option("Blue"),
+        ],
+    )
+    page.add(color_dropdown, submit_btn, output_text)
